Basic Refactor of file layout using Express + .pug
Sudo merge of the previous state to using Express framwork. Still has
old files that can need to be completely refactored or merged to new
Express state.

Main entry point is /app.js and landing page ("/") is routed using
Express to ./routes/index.js

./routes/index.js handles the HTTP/GET and HTTP/POST requests and routes
them the proper controller. Controller files are stubs for the time
being but this is where we would map to business logic.

./controllers/landing.js asks for the server to render (create HTML
response file) from the specified file in the ./view directory.